requirements.txt
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
bash==0.6
certifi==2021.5.30
charset-normalizer==2.0.3
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.8
django-mssql-backend==2.8.1
django-webpack-loader==1.1.0
djangorestframework==3.12.4
env==0.1.0
filelock==3.0.12
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
Markdown==3.3.4
mssql==1.0.1
mssql-django==1.0
mysql-connector-python==8.0.25
mysqlclient==2.0.3
node==0.9.26
odict==1.7.0
optional-django==0.3.0
platformdirs==2.0.2
plumber==1.6
psycopg2==2.9.1
pyodbc==4.0.32
python-dateutil==1.5
pytz==2021.1
react==4.3.0
requests==2.26.0
six==1.16.0
sql-server.pyodbc==1.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.6.0
webpack==6.0.0
whitenoise==5.3.0
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3